+++ src/apis/queryFunctions/category.ts
@@ -1,5 +1,9 @@
-import { category } from '@/mocks/data/category';
+import devApiClient from '@/apis/queryFunctions/devApiClient';
+import { Category } from '@/apis/types/category';
+import { Response } from '@/apis/types/common';
 
-export const getCategory = () => {
-  return category;
+export const getCategory = async () => {
+  const response = await devApiClient.get<Response<Category[]>>(`/v2/categories`);
+
+  return response.data.result_data;
 };
