+++ src/apis/types/category.ts
@@ -1,0 +1,6 @@
+export interface Category {
+  category_id: number;
+  name: string;
+  parent_id: number | null;
+  sub_categories: Category[];
+}
